refactor(alerter): extract alert_queries.go with scanAlert helper
Move 14 alert lifecycle functions into alert_queries.go. Extract
the duplicated 21-field alert scan pattern into scanAlert().

Part of #75.
